refactor: OAuth, Login API 호출 함수 분리
